+++ platform-operator/controllers/verrazzano/component/networkpolicies/netpol_values.go
@@ -28,8 +28,8 @@ type chartValues struct {
 	Rancher             *rancherValues           `json:"rancher,omitempty"`
 	Velero              *veleroValues            `json:"velero,omitempty"`
 	ArgoCD              *argoCDValues            `json:"argoCd,omitempty"`
-	Capi                *capiValues              `json:"capi,omitempty"`
 	FluentOperator      *fluentOperatorValues    `json:"fluentOperator,omitempty"`
+	ClusterAPI          *clusterAPIValues        `json:"clusterAPI,omitempty"`
 }
 
 type authproxyValues struct {
@@ -110,7 +110,7 @@ type argoCDValues struct {
 	Enabled bool `json:"enabled"` // Always write
 }
 
-type capiValues struct {
+type clusterAPIValues struct {
 	Enabled bool `json:"enabled"` // Always write
 }