+++ CHANGES.md
@@ -5,7 +5,7 @@ Features:
 - Added a None profile that comes with all components disabled by default
 - Added Thanos, which supports high availability and long-term storage on top of Prometheus
 - Enhanced the Verrazzano CLI to support installing and upgrading from a private registry and in air-gapped environments
-- Added support for Kubernetes v1.25 and v1.26, and dropped support of Kubernetes v1.21
+- Added support for Kubernetes v1.25 and v1.26, and dropped support of Kubernetes v1.21, v1.22 and v1.23
 - Added Fluent Operator to allow use of Fluent Bit for log processing, but it will be disabled by default
 
 Component version updates:
@@ -16,11 +16,14 @@ Component version updates:
 - Jaeger v1.42.0
 - Prometheus Operator v0.64.1
 - Prometheus v2.44.0
+- kube-state-metrics v2.8.2
 - kube-prometheus-stack Helm chart v45.25.0
+- kube-state-metrics Helm chart v5.6.4
 
 Components added:
 - Thanos v0.30.2
 - Fluent Operator v2.2.0
+- FluentbitOpensearchOutput 1.0.0
 
 Fixes: